+++ fabric/src/main/java/io/github/sakurawald/fuji/core/manager/impl/module/ModuleManager.java
@@ -163,7 +163,7 @@ public class ModuleManager extends BaseManager {
                 try {
                     initializer.doReload();
                 } catch (Exception e) {
-                    LogUtil.error("Failed to reload module.", e);
+                    LogUtil.error("Failed to reload module: initializer = {}", initializer.getClass().getName(), e);
                 }
             }
         );