+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/motd/MotdInitializer.java
@@ -65,7 +65,7 @@ public class MotdInitializer extends ModuleInitializer {
             byteArrayOutputStream = new ByteArrayOutputStream();
             ImageIO.write(bufferedImage, "PNG", byteArrayOutputStream);
         } catch (Exception e) {
-            LogUtil.error("Failed to read icon %s, did you forget to put the image file in proper location?".formatted(preferIcon), e);
+            LogUtil.error("Failed to read icon {}, did you forget to put the image file in proper location?", preferIcon, e);
             return Optional.empty();
         }