+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [fuji] fix: should open the GUI if the `About GUI` is opened by `Modules Inspection GUI`.
+- [core] feature: attach detailed exception message for error log.